requirements.txt
torch>=1.10.0
transformers>=4.30.0
Pillow>=9.0.0
opencv-python>=4.5.5
tqdm>=4.60.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.2
easyocr>=1.6.1
matplotlib>=3.4.3
pytesseract>=0.3.10
pdf2image>=1.16.3
requests>=2.26.0